Integration of ft into a working system.  * Made ftTransferModule compile.  * bugfixes to make ftserver1test work.  * New P3Pipe / P3Hub ...  * Added Test Notes.  * First functions added to ftcontroller  * added isOnline to p3ConnectMgr.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@698 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/ft/fttransfermodule.h
@@ -62,9 +62,9 @@ class peerInfo
 	public:
   std::string peerId;
   uint32_t state;
-  uint32_t desiredRate;
-  uint32_t actualRate;
-  
+  double desiredRate;
+  double actualRate;
+
   //current file data request
   uint64_t offset;
   uint32_t chunkSize;
@@ -122,8 +122,8 @@ private:
   std::map<std::string,peerInfo> mOnlinePeers;
   	
   bool     mFlag;  //1:transfer complete, 0: not complete
-  uint32_t desiredRate;
-  uint32_t actualRate;
+  double desiredRate;
+  double actualRate;
 };
 
 #endif  //FT_TRANSFER_MODULE_HEADER
